Update technology select, creation button. + more
Add a right hand Drawer to the charity projects list view that will be
used for searching and filtering.

Update technology select such that a single select supports selecting
more than one technology. This select will in future be used for
filtering as well as its current use for assigning tech to charity
projects.

Update creation button to be a FAB and place it at the top of the
Charity Projects view next to the title.